rename and cleanup rename mijson package to migraph
